*Training on stragglers results in better generalisation*
First I identified stragglers of each class manifold using the method from (https://www.nature.com/articles/s42256-023-00772-9). A total of 6599 stragglers and 63401 non-stragglers were found within the MNIST dataset. Next I trained two models with the same architecture. First one was trained on stragglers and evaluated on non-stragglers, and the other was trained on non-stragglers and evaluated on non-stragglers. I repeated the same experiment 5 times. The models achieved the train accuracy of 99.67% ± 0.1 when training on stragglers, and test accuracy of 56.23% ± 0.8. On the other hand, when trained on non-stragglers it achieved train accuracy of 99.98% ± 0.005 and test accuracy of 33.58% ± 1.1.