Finished one of the TODOs: "check how many noisy-labels straggler-based approach identified...". I have also fixed a minor issue in the method that added noise (before the modifications were in-place which resulted in adding more noise as the addition was done class-by-class).
This verified that the stragglers approach is truly the best when it comes to identifying label noise. When performing experiments on the side (test2.py) I found that the performance of confidence-based approach (when identifying label noise) is heavily reliable on the model used and training regime. When training for more epochs the precision and recall SIGNIFICANTLY increase (the figures added show this for differing number of epochs used for training).